+++ apps/toys/package.json
@@ -12,54 +12,54 @@
     "*": "prettier --write --ignore-unknown"
   },
   "dependencies": {
-    "@ban12/tailwindcss-safe-area": "^0.2.1",
+    "@ban12/tailwindcss-safe-area": "^0.2.3",
     "@darkroom.engineering/tempus": "^0.0.46",
     "@gsap/react": "^2.1.1",
-    "@hookform/resolvers": "^3.7.0",
+    "@hookform/resolvers": "^3.9.1",
     "@next/third-parties": "catalog:next15",
-    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-navigation-menu": "^1.2.1",
     "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-toggle-group": "^1.0.4",
-    "@radix-ui/react-tooltip": "^1.1.3",
+    "@radix-ui/react-toggle-group": "^1.1.0",
+    "@radix-ui/react-tooltip": "^1.1.4",
     "@repo/i18n": "workspace:*",
     "@repo/ui": "workspace:^",
-    "@simplewebauthn/server": "^10.0.0",
-    "ahooks": "^3.7.8",
+    "@simplewebauthn/server": "^11.0.0",
+    "ahooks": "^3.8.1",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.0",
-    "gsap": "^3.12.4",
+    "clsx": "^2.1.1",
+    "gsap": "^3.12.5",
     "intl-messageformat": "catalog:",
-    "lenis": "^1.1.5",
+    "lenis": "^1.1.16",
     "lucide-react": "catalog:",
     "next": "catalog:next15",
     "next-themes": "catalog:",
     "react": "catalog:react19",
     "react-dom": "catalog:react19",
     "react-hook-form": "catalog:",
-    "react-virtualized-auto-sizer": "^1.0.20",
+    "react-virtualized-auto-sizer": "^1.0.24",
     "react-window": "^1.8.10",
     "server-only": "^0.0.1",
-    "vaul": "^1.1.0",
-    "zod": "^3.22.4",
-    "zustand": "^5.0.0"
+    "vaul": "^1.1.1",
+    "zod": "^3.23.8",
+    "zustand": "^5.0.1"
   },
   "devDependencies": {
     "@repo/eslint-config": "workspace:^",
     "@repo/tailwind-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@types/lodash": "^4.17.12",
-    "@types/node": "^22.7.8",
-    "@types/react": "catalog:react19",
-    "@types/react-dom": "catalog:react19",
+    "@types/lodash": "^4.17.13",
+    "@types/node": "^22.9.0",
+    "@types/react": "npm:types-react@19.0.0-rc.1",
+    "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
     "@types/react-window": "^1.8.8",
     "@types/wicg-file-system-access": "^2023.10.5",
-    "autoprefixer": "^10.4.16",
+    "autoprefixer": "^10.4.20",
     "babel-plugin-react-compiler": "catalog:",
     "eslint": "catalog:eslint9",
-    "postcss": "^8.4.39",
+    "postcss": "^8.4.49",
     "tailwind-merge": "catalog:",
     "tailwindcss": "catalog:",
     "typescript": "catalog:"
