+++ pnpm-workspace.yaml
@@ -4,16 +4,16 @@ packages:
 
 catalog:
   "@types/node": "^22.9.0"
-  "postcss": "^8.4.48"
+  "postcss": "^8.4.49"
   "tailwind-merge": "^2.5.4"
-  "tailwindcss": "^3.4.14"
+  "tailwindcss": "^3.4.15"
   "typescript": "^5.6.3"
   "intl-messageformat": "^10.7.6"
   "next-themes": "^0.4.3"
   "react-hook-form": "^7.53.2"
   "lucide-react": "^0.456.0"
   "gsap": "^3.12.5"
-  "babel-plugin-react-compiler": "19.0.0-beta-63b359f-20241101"
+  "babel-plugin-react-compiler": "19.0.0-beta-a7bf2bd-20241110"
 
 catalogs:
   # Can be referenced through "catalog:react18"
@@ -48,5 +48,5 @@ catalogs:
     "eslint": "^9.14.0"
 
   turbo:
-    "turbo": "^2.2.3"
-    "eslint-config-turbo": "^2.2.3"
+    "turbo": "^2.3.0"
+    "eslint-config-turbo": "^2.3.0"
